Rename file DashboardSettings to DashboardSettingsPage.
