+++ components/dashboards-web-component/src/designer/DashboardSettingsPage.jsx
